Refactor Message UI components into dedicated module
Moved Message, MessageView, and related types from Controls to a new Message module. Split Message.Level, MessageView, and Presenter into separate files for better organization and maintainability. No functional changes; this improves code structure and clarity.
